+++ src/openms/source/MATH/STATISTICS/GammaDistributionFitter.cpp
@@ -70,8 +70,10 @@ namespace OpenMS
       int inputs() const { return m_inputs; }
       int values() const { return m_values; }
 
-      GammaFunctor(unsigned dimensions, const std::vector<DPosition<2> >* data)
-      : m_inputs(dimensions), m_values(data->size()), m_data(data) {}
+      GammaFunctor(unsigned dimensions, const std::vector<DPosition<2> >* data) :
+        m_inputs(dimensions), 
+        m_values(static_cast<int>(data->size())), 
+        m_data(data) {}
 
       int operator()(const Eigen::VectorXd &x, Eigen::VectorXd &fvec)
       {
